QSimd: fix MSVC C4319 warning on ARM
Says MSVC:

   qsimd.cpp(629): warning C4319: '~': zero extending 'QCpuFeatureType' to 'quint64' of greater size

On non-x86 architectures, QCpuFeatureType is just uint32_t, thus
the warning. Fix by casting to quint64 before negating.

Amends 5f7e02efb8b70adf00a4f393e7c2a1c93daef9b3.

+++ src/corelib/global/qsimd.cpp
@@ -626,7 +626,7 @@ uint64_t QT_MANGLE_NAMESPACE(qDetectCpuFeatures)()
     bool runningOnValgrind = false;
 #endif
     if (Q_UNLIKELY(!runningOnValgrind && minFeatureTest != 0 && (f & minFeatureTest) != minFeatureTest)) {
-        quint64 missing = minFeatureTest & ~f;
+        quint64 missing = minFeatureTest & ~quint64(f);
         fprintf(stderr, "Incompatible processor. This Qt build requires the following features:\n   ");
         for (uint i = 0; i < std::size(features_indices); ++i) {
             if (missing & (Q_UINT64_C(1) << i))
